move editinplace controller https://online.sbis.ru/opendoc.html?guid=4302562e-9843-4736-9903-91621628886f
