+++ tests/ControlsUnit/display/controllers/EditInPlace.test.ts
@@ -1,6 +1,6 @@
 import { assert } from 'chai';
 
-import * as EditInPlaceController from 'Controls/_display/controllers/EditInPlace';
+import * as EditInPlaceController from 'Controls/_editInPlace/EditInPlace';
 import AddInPlaceStrategy from 'Controls/_display/itemsStrategy/AddInPlace';
 
 describe('Controls/_display/controllers/EditInPlace', () => {
